+++ lua/snippets/luasnip/lua_snippets.lua
@@ -1,26 +1,40 @@
 local ls = require('luasnip')
 local fmta = require('luasnip.extras.fmt').fmta
 
-local quick_pack = ls.s(
-	{ trig = 'qp', name = 'New Packer Module', dscr = 'Creates a blank packer module' },
+local s = ls.snippet
+local isn = ls.indent_snippet_node
+local t = ls.text_node
+local i = ls.insert_node
+local c = ls.choice_node
+
+local packmod = s(
+	{ trig = 'packmod', name = 'New Packer Module', dscr = 'Creates a blank packer module' },
 	fmta(
 		[[
 		return function(use)
 			use({
-				'<>',
-				config = function()
-					<>
-				end,
+				'<repo>',
+				<config>
 			})
 		end
 		]],
 		{
-			ls.i(1, 'github_repo'),
-			ls.i(2),
+			repo = i(1, 'repo'),
+			config = isn(
+				2,
+				c(1, {
+					fmta('config = function ()\n\t<body>\nend,', { body = i(1) }),
+					t(''),
+				}),
+				'$PARENT_INDENT\t\t'
+			),
 		}
 	)
 )
 
 ls.add_snippets('lua', {
-	quick_pack,
+	packmod,
+	s('isn2', {
+		isn(1, ls.t({ 'This is', 'A multiline', 'comment' }), '$PARENT_INDENT\t'),
+	}),
 })