+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 > Bug Fixes
 - **luasnip:** update configuration
+- **snippet:** fixed lua snipped for pack module
 - **snippter:** apm:span func query capture fix
 
 > Code Refactoring